# backend/src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: Worship Core Backend
  version: 0.1.0
  description: Server-side session authentication; mutations require the CSRF header obtained from /api/auth/csrf.
paths:
...
  /api/auth/signup:
    post:
      operationId: signup
      responses:
        '200':
          description: Account created and verification delivery status returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityServiceDeliveryResultResponse'
        '400':
          $ref: '#/components/responses/Error400'
          description: Invalid email or password
        '401':
          $ref: '#/components/responses/Error401'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '403':
          $ref: '#/components/responses/Error403'
          description: Invalid CSRF token
        '404':
          $ref: '#/components/responses/Error404'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '405':
          $ref: '#/components/responses/Error405'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '406':
          $ref: '#/components/responses/Error406'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '409':
          $ref: '#/components/responses/Error409'
          description: Email already verified by another account
        '415':
          $ref: '#/components/responses/Error415'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '500':
          $ref: '#/components/responses/Error500'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '503':
          $ref: '#/components/responses/Error503'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
      security:
      - csrf: []
      x-csrf-header-source: GET /api/auth/csrf → headerName/token; refresh after login or session rotation
      parameters: []
      x-business-constraints: New password length 12–128; current-password verification only checks credential match. Email
        is canonicalized trim/lowercase; new verified ownership is unique.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/AccountControllerCredentialsRequest'
              - type: object
                properties:
                  password:
                    type: string
                    minLength: 12
                    maxLength: 128
  /api/auth/login:
    post:
      operationId: passwordLogin
      responses:
        '204':
          description: Authenticated JDBC session; refresh CSRF token after login
        '400':
          $ref: '#/components/responses/Error400'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '401':
          $ref: '#/components/responses/Error401'
          description: Authentication failed
        '403':
          $ref: '#/components/responses/Error403'
          description: Invalid CSRF token
        '404':
          $ref: '#/components/responses/Error404'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '405':
          $ref: '#/components/responses/Error405'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '406':
          $ref: '#/components/responses/Error406'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '409':
          $ref: '#/components/responses/Error409'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '415':
          $ref: '#/components/responses/Error415'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '500':
          $ref: '#/components/responses/Error500'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '503':
          $ref: '#/components/responses/Error503'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
      security:
      - csrf: []
      x-csrf-header-source: GET /api/auth/csrf → headerName/token; refresh after login or session rotation
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountControllerCredentialsRequest'
  /api/auth/verify-email:
    post:
      operationId: verifyEmail
      responses:
        '204':
          description: Email verified
        '400':
          $ref: '#/components/responses/Error400'
          description: Invalid expired or used token
        '401':
          $ref: '#/components/responses/Error401'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '403':
          $ref: '#/components/responses/Error403'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '404':
          $ref: '#/components/responses/Error404'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '405':
          $ref: '#/components/responses/Error405'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '406':
          $ref: '#/components/responses/Error406'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '409':
          $ref: '#/components/responses/Error409'
          description: Verified email already owned
        '415':
          $ref: '#/components/responses/Error415'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '500':
          $ref: '#/components/responses/Error500'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '503':
          $ref: '#/components/responses/Error503'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
      security:
      - csrf: []
      x-csrf-header-source: GET /api/auth/csrf → headerName/token; refresh after login or session rotation
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountControllerTokenRequestRequest'
  /api/auth/resend-verification:
    post:
      operationId: resendSignupVerification
      responses:
        '202':
          description: Request accepted without disclosing account state
        '400':
          $ref: '#/components/responses/Error400'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '401':
          $ref: '#/components/responses/Error401'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '403':
          $ref: '#/components/responses/Error403'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '404':
          $ref: '#/components/responses/Error404'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '405':
          $ref: '#/components/responses/Error405'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '406':
          $ref: '#/components/responses/Error406'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '409':
          $ref: '#/components/responses/Error409'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '415':
          $ref: '#/components/responses/Error415'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '500':
          $ref: '#/components/responses/Error500'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '503':
          $ref: '#/components/responses/Error503'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
      security:
      - csrf: []
      x-csrf-header-source: GET /api/auth/csrf → headerName/token; refresh after login or session rotation
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountControllerSignupResendRequest'
  /api/auth/request-password-reset:
    post:
      operationId: requestPasswordReset
      responses:
        '202':
          description: Request accepted without disclosing account state
        '400':
          $ref: '#/components/responses/Error400'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '401':
          $ref: '#/components/responses/Error401'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '403':
          $ref: '#/components/responses/Error403'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '404':
          $ref: '#/components/responses/Error404'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '405':
          $ref: '#/components/responses/Error405'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '406':
          $ref: '#/components/responses/Error406'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '409':
          $ref: '#/components/responses/Error409'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '415':
          $ref: '#/components/responses/Error415'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '500':
          $ref: '#/components/responses/Error500'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '503':
          $ref: '#/components/responses/Error503'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
      security:
      - csrf: []
      x-csrf-header-source: GET /api/auth/csrf → headerName/token; refresh after login or session rotation
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountControllerEmailRequestRequest'
  /api/auth/reset-password:
    post:
      operationId: resetPassword
      responses:
        '204':
          description: Password reset and all sessions invalidated
        '400':
          $ref: '#/components/responses/Error400'
          description: Invalid expired or used token
        '401':
          $ref: '#/components/responses/Error401'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '403':
          $ref: '#/components/responses/Error403'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '404':
          $ref: '#/components/responses/Error404'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '405':
          $ref: '#/components/responses/Error405'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '406':
          $ref: '#/components/responses/Error406'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '409':
          $ref: '#/components/responses/Error409'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '415':
          $ref: '#/components/responses/Error415'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '500':
          $ref: '#/components/responses/Error500'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '503':
          $ref: '#/components/responses/Error503'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
      security:
      - csrf: []
      x-csrf-header-source: GET /api/auth/csrf → headerName/token; refresh after login or session rotation
      parameters: []
      x-business-constraints: New password length 12–128; current-password verification only checks credential match. Email
        is canonicalized trim/lowercase; new verified ownership is unique.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/AccountControllerResetRequestRequest'
              - type: object
                properties:
                  password:
                    type: string
                    minLength: 12
                    maxLength: 128
  /api/account:
    get:
      operationId: getAccount
      security:
      - session: []
      responses:
        '200':
          description: Account DTO with email and login method metadata only
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityServiceAccountViewResponse'
        '400':
          $ref: '#/components/responses/Error400'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '401':
          $ref: '#/components/responses/Error401'
          description: Authentication required
        '403':
          $ref: '#/components/responses/Error403'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '404':
          $ref: '#/components/responses/Error404'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '405':
          $ref: '#/components/responses/Error405'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '406':
          $ref: '#/components/responses/Error406'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '409':
          $ref: '#/components/responses/Error409'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '415':
          $ref: '#/components/responses/Error415'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '500':
          $ref: '#/components/responses/Error500'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '503':
          $ref: '#/components/responses/Error503'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
      parameters: []
    delete:
      operationId: withdrawAccount
      security:
      - session: []
        csrf: []
      responses:
        '204':
          description: Account withdrawn and sessions revoked
        '400':
          $ref: '#/components/responses/Error400'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '401':
          $ref: '#/components/responses/Error401'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '403':
          $ref: '#/components/responses/Error403'
          description: Recent reauthentication or CSRF required
        '404':
          $ref: '#/components/responses/Error404'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '405':
          $ref: '#/components/responses/Error405'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '406':
          $ref: '#/components/responses/Error406'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '409':
          $ref: '#/components/responses/Error409'
          description: Workspace or document responsibility must be transferred first
        '415':
          $ref: '#/components/responses/Error415'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '500':
          $ref: '#/components/responses/Error500'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '503':
          $ref: '#/components/responses/Error503'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
      x-csrf-header-source: GET /api/auth/csrf → headerName/token; refresh after login or session rotation
      parameters: []
  /api/account/reauthenticate:
    post:
      operationId: passwordReauthenticate
      security:
      - session: []
        csrf: []
      responses:
        '204':
          description: Recent reauthentication recorded; refresh CSRF token
        '400':
          $ref: '#/components/responses/Error400'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '401':
          $ref: '#/components/responses/Error401'
          description: Password authentication failed
        '403':
          $ref: '#/components/responses/Error403'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '404':
          $ref: '#/components/responses/Error404'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '405':
          $ref: '#/components/responses/Error405'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '406':
          $ref: '#/components/responses/Error406'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '409':
          $ref: '#/components/responses/Error409'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '415':
          $ref: '#/components/responses/Error415'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '500':
          $ref: '#/components/responses/Error500'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '503':
          $ref: '#/components/responses/Error503'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
      x-csrf-header-source: GET /api/auth/csrf → headerName/token; refresh after login or session rotation
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountControllerPasswordRequestRequest'
  /api/account/emails:
    post:
      operationId: addUserEmail
      security:
      - session: []
        csrf: []
      responses:
        '200':
          description: Pending email created and delivery status returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IdentityServiceDeliveryResultResponse'
        '400':
          $ref: '#/components/responses/Error400'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '401':
          $ref: '#/components/responses/Error401'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '403':
          $ref: '#/components/responses/Error403'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '404':
          $ref: '#/components/responses/Error404'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '405':
          $ref: '#/components/responses/Error405'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '406':
          $ref: '#/components/responses/Error406'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '409':
          $ref: '#/components/responses/Error409'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '415':
          $ref: '#/components/responses/Error415'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '500':
          $ref: '#/components/responses/Error500'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '503':
          $ref: '#/components/responses/Error503'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
      x-csrf-header-source: GET /api/auth/csrf → headerName/token; refresh after login or session rotation
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountControllerEmailRequestRequest'
  /api/account/emails/{emailId}:
    delete:
      operationId: removeUserEmail
      security:
      - session: []
        csrf: []
      responses:
        '204':
          description: Email removed
        '400':
          $ref: '#/components/responses/Error400'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '401':
          $ref: '#/components/responses/Error401'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '403':
          $ref: '#/components/responses/Error403'
          description: Recent reauthentication required
        '404':
          $ref: '#/components/responses/Error404'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '405':
          $ref: '#/components/responses/Error405'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '406':
          $ref: '#/components/responses/Error406'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '409':
          $ref: '#/components/responses/Error409'
          description: Last usable login method must remain
        '415':
          $ref: '#/components/responses/Error415'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '500':
          $ref: '#/components/responses/Error500'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
        '503':
          $ref: '#/components/responses/Error503'
          description: Standard capability/protocol error; see ApiError.code and current authorization/provider configuration
      x-csrf-header-source: GET /api/auth/csrf → headerName/token; refresh after login or session rotation
      parameters:
      - name: emailId
        in: path
        required: true
        schema:
          type: integer
          format: int64
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: SESSION
    csrf:
      type: apiKey
      in: header
      name: X-CSRF-TOKEN
      description: Use actual headerName/token from /api/auth/csrf, not a long-lived cached token
  schemas:
...
    IdentityServiceDeliveryResultResponse:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        deliveryStatus:
          type: string
      required:
      - userId
      - deliveryStatus
...
  responses:
    Error400:
      description: Standard error envelope
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiErrorResponse'
            - type: object
              properties:
                code:
                  type: string
                  enum:
                  - VALIDATION_FAILED
    Error401:
      description: Standard error envelope
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiErrorResponse'
            - type: object
              properties:
                code:
                  type: string
                  enum:
                  - AUTHENTICATION_REQUIRED
    Error403:
      description: Standard error envelope
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiErrorResponse'
            - type: object
              properties:
                code:
                  type: string
                  enum:
                  - ACCESS_DENIED
                  - REAUTHENTICATION_REQUIRED
                  - YOUTUBE_AUTHORIZATION_REQUIRED
    Error404:
      description: Standard error envelope
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiErrorResponse'
            - type: object
              properties:
                code:
                  type: string
                  enum:
                  - NOT_FOUND
                  - REQUEST_REJECTED
    Error405:
      description: Standard error envelope
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiErrorResponse'
            - type: object
              properties:
                code:
                  type: string
                  enum:
                  - REQUEST_REJECTED
    Error406:
      description: Standard error envelope
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiErrorResponse'
            - type: object
              properties:
                code:
                  type: string
                  enum:
                  - REQUEST_REJECTED
    Error409:
      description: Standard error envelope
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiErrorResponse'
            - type: object
              properties:
                code:
                  type: string
                  enum:
                  - STATE_CONFLICT
                  - VERSION_CONFLICT
                  - CONSTRAINT_CONFLICT
    Error415:
      description: Standard error envelope
      content:
        application/json:
          schema:
            allOf:
            - $ref: '#/components/schemas/ApiErrorResponse'
            - type: object
              properties:
                code:
                  type: string
                  enum:
                  - REQUEST_REJECTED
...